tested general reproducibility and fixed a resulting bug
